+++ CHANGELOG.md
@@ -78,6 +78,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Send Integration ID to SAP for detector creation [(#712)](https://github.com/wazuh/wazuh-indexer-plugins/pull/712)
 - Update Content Manager docs [(#716)](https://github.com/wazuh/wazuh-indexer-plugins/pull/716)
 - Update CTI field mappings [(#718)](https://github.com/wazuh/wazuh-indexer-plugins/pull/718)
+- Update content spaces [(#724)](https://github.com/wazuh/wazuh-indexer-plugins/pull/724)
 - Improve Content Manager code quality [(#720)](https://github.com/wazuh/wazuh-indexer-plugins/pull/720)
 
 ### Deprecated